requirements.txt
gcloud==0.18.3
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-core==2.0.0
google-cloud-translate==3.3.2
oauth2client==4.1.3